--- src/main/java/kitchenpos/menugroup/MenuGroupMapper.java
@@ -1,21 +1,0 @@
-package kitchenpos.menugroup;
-
-public class MenuGroupMapper {
-
-    private MenuGroupMapper() {
-    }
-
-    public static MenuGroupDto toDto(MenuGroup menuGroup) {
-        MenuGroupDto menuGroupDto = new MenuGroupDto();
-        menuGroupDto.setId(menuGroup.getId());
-        menuGroupDto.setName(menuGroup.getName());
-        return menuGroupDto;
-    }
-
-    public static MenuGroup toEntity(MenuGroupDto menuGroupDto) {
-        return new MenuGroup.Builder()
-            .id(menuGroupDto.getId())
-            .name(menuGroupDto.getName())
-            .build();
-    }
-}
