+++ src/main/java/kitchenpos/menugroup/service/MenuGroupMapper.java
@@ -1,0 +1,15 @@
+package kitchenpos.menugroup.service;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MenuGroupMapper {
+
+    public MenuGroup toEntity(MenuGroupDto menuGroupDto) {
+        return new MenuGroup.Builder()
+            .id(menuGroupDto.getId())
+            .name(menuGroupDto.getName())
+            .build();
+    }
+}
