+++ src/main/java/kitchenpos/menugroup/service/MenuGroupService.java
@@ -1,6 +1,8 @@
-package kitchenpos.menugroup;
+package kitchenpos.menugroup.service;
 
 import java.util.List;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
